requirements.txt
flask
python-dotenv
openai
tiktoken
elasticsearch==7.12.0
langchain
langchain-openai 
langchain_community
langchain-elasticsearch